Move CategoryCreateOnRegisterView into CategoryComponent. Update DashboardComponent.kt for new UseCases & ViewModel
