+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/compose/dashboard/DashboardComponent.kt
@@ -1,9 +1,9 @@
 package de.hsfl.budgetBinder.compose.dashboard
 
 import androidx.compose.runtime.*
-import de.hsfl.budgetBinder.domain.use_case.get_user.UserUseCase
 import de.hsfl.budgetBinder.presentation.Screen
-import de.hsfl.budgetBinder.presentation.UserViewModel
+import de.hsfl.budgetBinder.domain.usecase.*
+import de.hsfl.budgetBinder.presentation.viewmodel.DashboardViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -14,13 +14,15 @@ import org.kodein.di.instance
 fun DashboardComponent(screenState: MutableState<Screen>) {
     val scope = CoroutineScope(Dispatchers.Unconfined + SupervisorJob())
     val di = localDI()
-    val userUseCase: UserUseCase by di.instance()
-    val userViewModel = UserViewModel(userUseCase, scope)
-    val viewState = userViewModel.state.collectAsState(scope)
+    val getAllEntriesUseCase: GetAllEntriesUseCase by di.instance()
+    val getAllCategoriesUseCase: GetAllCategoriesUseCase by di.instance()
+    val logoutUseCase : LogoutUseCase by di.instance()
+    val dashboardViewModel = DashboardViewModel(getAllEntriesUseCase,getAllCategoriesUseCase,logoutUseCase, scope)
+    val viewState = dashboardViewModel.state.collectAsState(scope)
 
     DashboardView(
         state = viewState,
-        onUpdate = { userViewModel.getMyUser() },
+        onUpdate = { },
         onCategorySummaryButton = { screenState.value = Screen.CategorySummary},
         onSettingsButton = {screenState.value = Screen.Settings},
         onEntryCreateButton = {screenState.value = Screen.EntryCreate},
